+++ ci/do_ci.sh
@@ -71,6 +71,8 @@ echo "make command: ${MAKE_COMMAND}"
 
 export BAZEL_CXXOPTS="-std=c++17"
 
+# Work around for https://github.com/actions/runner-images/issues/13564
+export USE_BAZEL_VERSION="8.5.0"
 
 BAZEL_OPTIONS_DEFAULT="--copt=-DENABLE_METRICS_EXEMPLAR_PREVIEW --//exporters/otlp:with_otlp_grpc_credential_preview=true"
 BAZEL_OPTIONS="$BAZEL_OPTIONS_DEFAULT"
@@ -104,6 +106,12 @@ CMAKE_OPTIONS+=("-DCMAKE_CXX_EXTENSIONS=OFF")
 if [ -n "$CMAKE_TOOLCHAIN_FILE" ]; then
   echo "CMAKE_TOOLCHAIN_FILE is set to: $CMAKE_TOOLCHAIN_FILE"
   CMAKE_OPTIONS+=("-DCMAKE_TOOLCHAIN_FILE=$CMAKE_TOOLCHAIN_FILE")
+fi
+
+if [ -n "$OTELCPP_CMAKE_CACHE_FILE" ]; then
+  OTELCPP_CMAKE_CACHE_FILE_PATH="${SRC_DIR}/test_common/cmake/${OTELCPP_CMAKE_CACHE_FILE}"
+else
+  OTELCPP_CMAKE_CACHE_FILE_PATH="${SRC_DIR}/test_common/cmake/all-options-abiv1-preview.cmake"
 fi
 
 echo "CMAKE_OPTIONS:" "${CMAKE_OPTIONS[@]}"
@@ -339,19 +347,26 @@ elif [[ "$1" == "cmake.legacy.test" ]]; then
   make test
   exit 0
 elif [[ "$1" == "cmake.clang_tidy.test" ]]; then
-  cd "${BUILD_DIR}"
-  rm -rf *
-  export BUILD_ROOT="${BUILD_DIR}"
-  cmake -S ${SRC_DIR} \
+  rm -rf "${BUILD_DIR}"
+  mkdir -p "${BUILD_DIR}"
+  clang-tidy --version
+  LOG_FILE="${BUILD_DIR}/opentelemetry-cpp-clang-tidy.log"
+  cmake "${CMAKE_OPTIONS[@]}"  \
+    -S ${SRC_DIR} \
     -B ${BUILD_DIR} \
-    -C ${SRC_DIR}/test_common/cmake/all-options-abiv2-preview.cmake  \
-    "${CMAKE_OPTIONS[@]}" \
+    -C ${OTELCPP_CMAKE_CACHE_FILE_PATH} \
     -DWITH_OPENTRACING=OFF \
     -DCMAKE_CXX_FLAGS="-Wno-deprecated-declarations" \
     -DCMAKE_EXPORT_COMPILE_COMMANDS=ON \
-    -DCMAKE_CXX_CLANG_TIDY="clang-tidy;--quiet;-p;${BUILD_DIR}"
-  make -j $(nproc)
-  make test
+    -DCMAKE_CXX_CLANG_TIDY="clang-tidy;--header-filter=.*/opentelemetry-cpp/.*;--exclude-header-filter=.*(internal/absl|third_party|third-party|build.*|/usr|/opt)/.*|.*\.pb\.h;--quiet"
+  cmake --build "${BUILD_DIR}" -- -j $(nproc) 2>&1 | tee "$LOG_FILE"
+  if [ ! -s "$LOG_FILE" ]; then
+    echo "Error: Build log was not created at $LOG_FILE"
+    exit 1
+  fi
+  echo "Build log written to: $LOG_FILE"
+  echo "To generate a clang-tidy report, use the following command:"
+  echo "  python3 ./ci/create_clang_tidy_report.py --build_log $LOG_FILE"
   exit 0
 elif [[ "$1" == "cmake.legacy.exporter.otprotocol.test" ]]; then
   cd "${BUILD_DIR}"
@@ -492,7 +507,6 @@ elif [[ "$1" == "cmake.install.test" ]]; then
     "exporters_prometheus_builder"
     "exporters_elasticsearch"
     "exporters_zipkin"
-    "exporters_zipkin_builder"
   )
   EXPECTED_COMPONENTS_STRING=$(IFS=\;; echo "${EXPECTED_COMPONENTS[*]}")
   mkdir -p "${BUILD_DIR}/install_test"