+++ CHANGELOG.md
@@ -15,6 +15,42 @@ Increment the:
 
 ## [Unreleased]
 
+* [CONFIGURATION] File configuration - remove zipkin
+  [#3804](https://github.com/open-telemetry/opentelemetry-cpp/pull/3804)
+
+* [CONFIGURATION] File configuration - rename tls properties
+  [#3805](https://github.com/open-telemetry/opentelemetry-cpp/pull/3805)
+
+* [API] Parse baggage value as spec compliant
+  [#3758](https://github.com/open-telemetry/opentelemetry-cpp/pull/3758)
+
+* [CONFIGURATION] File configuration - implement log_level
+  [#3810](https://github.com/open-telemetry/opentelemetry-cpp/pull/3810)
+
+* [SEMANTIC CONVENTIONS] Upgrade to semantic conventions 1.39.0
+  [#3813](https://github.com/open-telemetry/opentelemetry-cpp/pull/3813)
+
+* [CONFIGURATION] File configuration - prometheus without_target_info
+  [#3818](https://github.com/open-telemetry/opentelemetry-cpp/pull/3818)
+
+* [BAZEL] Fix CI after bazel 9.0.0 release
+  [#3823](https://github.com/open-telemetry/opentelemetry-cpp/pull/3823)
+
+* [CONFIGURATION] File configuration - implement cardinality_limits
+  [#3820](https://github.com/open-telemetry/opentelemetry-cpp/pull/3820)
+
+Breaking changes:
+
+* [CONFIGURATION] File configuration - remove zipkin
+  [#3804](https://github.com/open-telemetry/opentelemetry-cpp/pull/3804)
+  * The zipkin trace exporter is no longer supported
+    by declarative configuration, due to an upstream schema change.
+
+* [CONFIGURATION] File configuration - rename tls properties
+  [#3805](https://github.com/open-telemetry/opentelemetry-cpp/pull/3805)
+  * The Tls properties for Grpc and Http are renamed,
+    due to an upstream schema change.
+
 ## [1.24 2025-11-20]
 
 * [RELEASE] Bump main branch to 1.24-dev